Added support for vertical scrolling of the line graph parent view, even in cases where the touch starts inside the graph view. This makes it possible to have several graph views within a UIScrollView.
To achieve this I needed to expose the panView and the panGesture as properties and add the method gestureRecognizerShouldBegin, which checks the direction of the pan and only starts the recogniser if the pan is in the horizontal direction. Vertical panning will instead be handled by the parent view (usually a UIScrollView).

— Nils