Added support for scrolling with a GraphView placed in UIScrollView
To enable scrolling with a touch that begins on a GraphView I made two changes, first I exposed panView and panGesture as properties. Second I implemented the gestureRecognizerShouldBegin method, and had it dodge panning in the vertical direction, instead sending it to the parent view which means that scrolling a UIScrollView will work as normal.

— Nils